+++ app/src/main/java/com/djacoronel/goalmentum/storage/model/Work.kt
@@ -17,7 +17,7 @@ class Work : BaseModel {
     var id: Long = 0
 
     @Column
-    var positionInList: Int = 0
+    var positionInList: Long = 0
 
     @Column
     var assignedMilestone: Long = 0
@@ -36,7 +36,7 @@ class Work : BaseModel {
 
     constructor()
 
-    constructor(id:Long, positionInList: Int, assignedMilestone: Long, description: String, date: Date, achieved: Boolean, dateAchieved: Date) {
+    constructor(id:Long, positionInList: Long, assignedMilestone: Long, description: String, date: Date, achieved: Boolean, dateAchieved: Date) {
         this.id = id
         this.positionInList = positionInList
         this.assignedMilestone = assignedMilestone
